+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 * Make unlimited positionals vs. unlimited options more intuitive [#102]
 * Add missing getters `get_options` and `get_description` to App [#105]
 * The app name now can be set, and will override the auto name if present [#105]
+* Add `(REQUIRED)` for required options [#104]
+* Print simple name for Needs/Excludes [#104]
+* Use Needs instead of Requires in help print [#104]
 
 [#102]: https://github.com/CLIUtils/CLI11/issues/102
 [#105]: https://github.com/CLIUtils/CLI11/issues/105
+[#104]: https://github.com/CLIUtils/CLI11/pull/104
 
 
 ## Version 1.5: Optionals